Add story to Table
